+++ schema/interfaces.ts
@@ -11,12 +11,12 @@ export default gql`
   }
 
   interface Attributes {
-    strength: Int!
-    dexterity: Int!
-    constitution: Int!
+    strength: Float!
+    dexterity: Float!
+    constitution: Float!
 
-    intelligence: Int!
-    wisdom: Int!
-    charisma: Int!
+    intelligence: Float!
+    wisdom: Float!
+    charisma: Float!
   }
 `;
